+++ app/javascript/components/dashboard/card/CardDetailsContainer.jsx
@@ -2,40 +2,21 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import CardLabels from './CardLabels';
-import moment from 'moment';
+import CardDueDate from './CardDueDate';
 
 class CardDetailsContainer extends React.Component {
   render() {
-    const store = this.context.store;
     const card = this.props.card;
 
-    const date = card.due_date ? moment(card.due_date).format('MMM D') : null;
-    let dueStatus;
-
-    if (date) {
-      const now = moment()
-      const diff = now.diff(moment(card.due_date), 'days')
-
-      if (diff < 0) {
-        dueStatus = 'completed';
-      } else if (diff > 0) {
-        dueStatus = 'overdue';
-      } else {
-        dueStatus = 'due-soon';
-      }
-    }
     return (
       <li className="details-section">
         <ul className="modal-details-list">
           <CardLabels
             card={card}
           />
-          <li className="due-date-section">
-            <h3>Due Date</h3>
-            <div id="dueDateDisplay" className="overdue completed">
-              <input id="dueDateCheckbox" type="checkbox" className="checkbox" checked="" />Aug 4 at 10:42 AM <span>(past due)</span>
-            </div>
-          </li>
+          <CardDueDate
+            card={card}
+          />
         </ul>
         <form className="description">
           <p>Description</p>
